Use engine thumbnail generator (#15)
JIRA: SMOK-49295

Replace code that generate thumbnails by code in the engine that does the same
thing. The code in the engine leverage new code starting in Flame 2019.1 that
use the Flame export engine instead of ffmpeg.

This however means that some of the heavy lifting done in the background job
needs to be moved to the Flame process in the python hooks. This is needed
since the Flame API only work in the Flame process.

Add Quicktime handler information to allow pre-import of quicktime file.

Upgrade to tk-flame v1.14.3